Handle duplications with family address locate on network interface
utlities with the family address on a network interface, implemented as nodes in a list, always involves finding the current address family(ipv4, or ipv6 or etc.) on the particular interface since netbsd uses per interface  address family. this locate routine has been duplicated throughout the whole API and must be handled for better maintainability.