remove dead error condition in marking congested experienced packet routine
this condition is automatically not met if the 'if (t != NULL)' condition takes a false branch.
if checking the non-nullity of the packet tag takes a true branch, it means that the altq tag pointer (at) can never be null
since it is initialized with the packet tag which is also not null, so 'if (at == NULL) remains to be false even if the outer if condition
takes a true branch. this makes that portion of the code dead and never evaluated.

the dead condition here is return 0 on failed marking, but the else statement in this particukar if chain handles that for ussince t being NULL and packettr
also being NULL will evaluate to failure of the marking hence agreeing with the else statement in this condition chain