handle network authorization to remove the presence of two switch statements in altqioctl
in altqioctl function, there are two switch statements. the first switch contains cases for ALTQGTYPE and ALTQTBRGET which breaks.
and sets a default case to return error if the altq network operation authorization fails. when ALTQTBRSET is executing, it goes to the default section in the first switch, because it isn't included in the non-default case block.
and if the authorization condition in the default block is true, it returns, and if it's false, it goes to the next switch statement and executes the setting of the token bucket regulator. same applies for all default statements. authorization condition of first switch is checked, if false, if moves to the default block of the second switch statement.

This is redundant and can be expressed using a single switch statement where ALTQGTYPE and ALTQBRGET remains the same because it's unaffected by the first. since only the ALTQBRSET and default is affected by the authorization check before executing, it can be packaged in a function and callled in the ALTQBRSET and default case of the second one direclt needing just one switch statement now!