Handle very large and messy blueioctl function block to improve readability
the blueioclt function contains enable, disable, if_attach, if_detach, get_stats, and config routines.
all these routines are choked in one function block which can be organized into functions that focus on
that one routine.
for if_attach, due to consistent check for error codes and early termination in the allocation of
the blue queue state, an extra ENOMEM error check has to be run after blue_state_alloc function
to rightly get the switch program to terminate early if any of the return values of error matches the ENOMEM error code.

for the error passed as parameters, had to pass them by pointers so that copies are not created after the function terminates but rather hold the error value in the
function error variable throughout.