revert set configuration and get stats back to original state
when blue is attached to an interface, the blue state is allocated on the heap in BLUE_IF_ATTACH case.
when we are getting stats and configurations, we will need to access that memeory address of the queue and manipulates its fields.

passing the blue_state rqp as parameter does not use the memory of the dynamically allocated blue state and hence poses a state of use of unitialized blue state.

we will have to allocate the heap out of a larger scope but that compromises safety in possibly missing freeing points.