handle blue_addq large function block by moving the force drop routine into a function
when blue is enqueueing a packet, it uses a drop/mark scheme based on probabilities.
in dropping, it either uses the drop early or force drop scheme based on the droptype scheme set(either early or forced.)
in the else statements that forced drops, should be put in a function block as it handles the large block of add queue and also sets it to focus on just one thing.
can also be reused anywhere in the code