make flowlist_lookup function more maintainable
during the search of the flowlist, it uses conditionals for ipv4 and ipv6 family addresses.
so each address family search is given that as a single responsiblity to a funcion.

in its original form, there are two breaks which after exiitng the switch statements, saves the current flow count value as the number of active fves.
a third conditioon which returns the fve found and does not proceed with the saving of the active number of fves.

in my funciton, a NULL returning fve indicates no termination by return statement hence flows value gets to be saved. a non-NULL fve indicates a termination by returning and hence no active flow value gets to be saved