Increase the number of CBQ and HFSC classes
very large networks are common these days and in some configs, we will need more classes
than ALTQ supports by default. increase the maximum number of classes we allow.
this is not an expensive increment as it will cost an almost neglible(trivial) amount of memory.
can make it tunable in the future if need arises