remove goto statment
in finding an underlimit class when dequeing packets, if no underlimit class is found
over a sinlge iteration, the cutof is increased and the routine is performed again and again
till an underlimit class is found.
instead of using a goto again lable, we can simply use a do while which will do the underlimit
class finding and before checking the while condition, checks if a class was found,
if not, increments the cutoff and try again based on the same condition that the cutoff was increased which is simply
an underlimit class not found.