remove goto statement
use do while loops in place of goto statement labels.. in looping through for a priority level,
the loop is run twice. the simplest is to send a loop_count and condition for twice running in the while condition.
and also increment the dificit to two for our second run in the loop.