check this out
so this is just using a flag switch so which you define is passed to the function that evaliuates the separate map segments

https://datatracker.ietf.org/doc/html/rfc6145

look at the abstract, I'm not really sure if its 100 percent right to use SIIT atm. but read it and see if they talk about IPV4 IPV6 translation.
if it does, learn the algorithm and use it to do the ipv6 to ipv4 tranlation.